backend/requirements.txt
asgiref
certifi
charset-normalizer
coreapi
coreschema
Django
django-cors-headers
django-filter
django-jsonfield
djangorestframework
djangorestframework-simplejwt
drf_spectacular
idna
inflection
itypes
Jinja2
MarkupSafe
openapi-codec
packaging
Pillow
PyJWT
pyparsing
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
simplejson
sqlparse
typing-extensions
uritemplate
urllib3